GetPreorderAccounts is somewhat working. Instead of using Date in CoffeeService as a parameter, we use String.
That commit message was wrong, it is GetPreOrderGoods not GetPreorderAccounts. We extracted it in a method which is called when FAB is clicked in Document fragment. This was a temporary solution, Document fragment is for documents, meaning this code should be moved to Fragment that deals with PreOrders. DocumentFragment should handle CoffeeDocs and DocGoods.